requirements.txt
opencv-python
face-recognition
numpy
ultralytics
pillow
requests